+++ src/deepl.ts
@@ -1,7 +1,16 @@
 import * as vscode from 'vscode';
+import * as debug from './debug';
 import { state } from './state';
 import { Language, SourceLanguageCode, TargetLanguageCode, TextResult, Translator } from 'deepl-node';
 import { EXTENSION_IDENTIFIER } from './constants';
+import { showSourceLanguagePrompt, showTargetLanguagePrompt } from './prompts';
+
+enum ResolveUnsuccessfulTranslationActions {
+  // eslint-disable-next-line @typescript-eslint/naming-convention
+  SELECT_SOURCE_LANGUAGE = 'Select source language and try again',
+  // eslint-disable-next-line @typescript-eslint/naming-convention
+  CHANGE_TARGET_LANGUAGE = 'Change target language and try again'
+}
 
 const cache = {
   targetLanguages: [] as readonly Language[],
@@ -22,12 +31,46 @@ const createTranslator = (apiKey: string) => {
   });
 };
 
-export function translate<T extends string | string[]>(texts: T, sourceLang: SourceLanguageCode | null, targetLang: TargetLanguageCode): Promise<T extends string ? TextResult : TextResult[]> {
+const handleTranslationFailure = async <T extends string | string[]>(texts: T, results: T extends string ? TextResult : TextResult[], sourceLanguage: SourceLanguageCode | undefined, targetLanguage: TargetLanguageCode, retries: number = 1) => {
+  debug.write('The translation result is the with the original input');
+
+  const selectedAction = await vscode.window.showWarningMessage(
+    'Translation was not successful!',
+    { 
+      detail: sourceLanguage
+        ? `Please check if you are using the correct source and target language. \n\n Source language: '${sourceLanguage}' \n Target language: '${targetLanguage}'`
+        : `It is possible that the source language could not be recognized correctly or the wrong target language has been selected. \n\n Target language: '${targetLanguage}'`,
+      modal: true
+    },
+    ResolveUnsuccessfulTranslationActions.SELECT_SOURCE_LANGUAGE,
+    ResolveUnsuccessfulTranslationActions.CHANGE_TARGET_LANGUAGE);
+
+  if (selectedAction === ResolveUnsuccessfulTranslationActions.SELECT_SOURCE_LANGUAGE) {
+    const newSourceLanguage = await showSourceLanguagePrompt();
+    if (newSourceLanguage) {
+      debug.write(`Retrying translation with new source language '${newSourceLanguage}'`);
+      return await translate(texts, newSourceLanguage, targetLanguage, retries - 1);
+    }
+  }
+
+  if (selectedAction === ResolveUnsuccessfulTranslationActions.CHANGE_TARGET_LANGUAGE) {
+    const newTargetLanguage = await showTargetLanguagePrompt();
+    if (newTargetLanguage) {
+      state.targetLanguage = newTargetLanguage;
+      debug.write(`Retrying translation with changed target language '${newTargetLanguage}'`);
+      return await translate(texts, sourceLanguage, newTargetLanguage, retries - 1);
+    }
+  }
+
+  return results;
+};
+
+export async function translate<T extends string | string[]>(texts: T, sourceLanguage: SourceLanguageCode | undefined, targetLanguage: TargetLanguageCode, retries: number = 1): Promise<T extends string ? TextResult : TextResult[]> {
   const translator = createTranslator(state.apiKey!);
-  return translator.translateText(
+  const results = await translator.translateText(
     texts,
-    sourceLang ?? null,
-    targetLang,
+    sourceLanguage ?? null,
+    targetLanguage,
     {
       formality: state.formality || undefined,
       glossary: state.glossaryId || undefined,
@@ -39,10 +82,19 @@ export function translate<T extends string | string[]>(texts: T, sourceLang: Sou
       tagHandling: state.tagHandling || undefined
     }
   );
+
+  const wasTranslationSuccessful = typeof texts === "string"
+    ? (results as TextResult).text !== texts
+    : (results as TextResult[]).filter((result, index) => result.text === texts[index]).length === 0;
+
+  return !wasTranslationSuccessful && retries > 0
+    ? handleTranslationFailure(texts, results, sourceLanguage, targetLanguage, retries)
+    : results;
 }
 
 export async function getTargetLanguages() {
   if (!state.apiKey) {
+    debug.write('Could not load target languages - api key is not configured!');
     return [];
   }
 
@@ -50,7 +102,7 @@ export async function getTargetLanguages() {
     return cache.targetLanguages;
   }
 
-  const translator = createTranslator(state.apiKey!);
+  const translator = createTranslator(state.apiKey);
   const languages = await translator.getTargetLanguages();
   cache.targetLanguages = languages;
   return languages;
@@ -58,6 +110,7 @@ export async function getTargetLanguages() {
 
 export async function getSourceLanguages() {
   if (!state.apiKey) {
+    debug.write('Could not load source languages - api key is not configured!');
     return [];
   }
 
@@ -65,7 +118,7 @@ export async function getSourceLanguages() {
     return cache.sourceLanguages;
   }
 
-  const translator = createTranslator(state.apiKey!);
+  const translator = createTranslator(state.apiKey);
   const languages = await translator.getSourceLanguages();
   cache.sourceLanguages = languages;
   return languages;