+++ src/status-bar.ts
@@ -2,7 +2,7 @@ import * as vscode from 'vscode';
 import * as deepl from './deepl';
 import { state } from './state';
 import { effect } from '@vue/reactivity';
-import { COMMAND_CONFIGURE, COMMAND_TRANSLATE_TO } from './constants';
+import { COMMAND_CONFIGURE, COMMAND_SET_TARGET_LANGAUGE } from './constants';
 
 export function createStatusBarItem() {
   const statusBarItem = vscode.window.createStatusBarItem(vscode.StatusBarAlignment.Left, 10);
@@ -17,19 +17,19 @@ export function createStatusBarItem() {
     }
 
     if (state.apiKey) {
-      statusBarItem.command = COMMAND_TRANSLATE_TO;
+      statusBarItem.command = COMMAND_SET_TARGET_LANGAUGE;
       statusBarItem.tooltip = 'Select the language you want to translate into';
     }
 
     if (state.apiKey && !state.targetLanguage) {
-      statusBarItem.command = COMMAND_TRANSLATE_TO;
+      statusBarItem.command = COMMAND_SET_TARGET_LANGAUGE;
       statusBarItem.text = prefix + 'Select language';
     }
 
     if (state.apiKey && state.targetLanguage) {
       const languages = await deepl.getTargetLanguages();
       const selectedLanguage = languages.find(x => x.code.toLocaleLowerCase() === state.targetLanguage!.toLocaleLowerCase());
-      statusBarItem.command = COMMAND_TRANSLATE_TO;
+      statusBarItem.command = COMMAND_SET_TARGET_LANGAUGE;
       statusBarItem.text = prefix + (selectedLanguage?.name ?? state.targetLanguage);
     }
   });